[FIR] Minor: Use separate `...Utils` object instead of a super class
In the following commits, `checkedOrigins` will be used by a checker
for equality bounds as well.
